requirements.txt

Flask==2.1.1
pytesseract==0.3.8
Pillow==9.0.1
gTTS==2.2.3
gingerit==0.8.0
